requirements.txt
PyQt6==6.6.1
markdown==3.5.2
PyQt6-QScintilla==2.14.1
pygments==2.17.2
watchdog==3.0.0
mistune==3.0.2 
pyinstaller
pillow